move TeacherInfoPage to new folder
